feat(Examples): create recorder example
