+++ examples/README.md
@@ -1,7 +1,8 @@
 # Examples
 
 | Example                  | Description                                                                                                                                                       |
-| ------------------------ | ----------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+|--------------------------|-------------------------------------------------------------------------------------------------------------------------------------------------------------------|
 | [Basic](./basic)         | A simple "Hello World" TypeScript example that plays an mp3 file. Notably, it works with discord.js v12 and so it also contains an example of creating an adapter |
 | [Radio Bot](./radio-bot) | A fun JavaScript example of what you can create using @discordjs/voice. A radio bot that plays output from your speakers in a Discord voice channel               |
 | [Music Bot](./music-bot) | A TypeScript example of a YouTube music bot. Demonstrates how queues can be implemented and how to implement "good" disconnect/reconnection logic                 |
+| [Recorder](./recorder)   | An example of using voice receive to create a bot that can record audio from users                                                                                |
