+++ examples/recorder/README.md
@@ -1,6 +1,7 @@
-# Receive Bot
+# 👂 Recorder Bot
 
-TODO
+This example shows how you can use the voice receive functionality in @discordjs/voice to record users in voice channels
+and save the audio to local Ogg files.
 
 ## Usage
 
@@ -10,7 +11,7 @@ $ npm install
 $ npm run build
 
 # Open this example and install dependencies
-$ cd examples/receiver
+$ cd examples/recorder
 $ npm install
 
 # Set a bot token (see auth.example.json)
@@ -19,7 +20,3 @@ $ nano auth.json
 # Start the bot!
 $ npm start
 ```
-
-## Code structure
-
-TODO
