+++ examples/recorder/auth.example.json
