+++ examples/recorder/src/createListeningStream.ts
